+++ src/strings/PowerFxResources.en-US.resx
@@ -2400,33 +2400,43 @@
   </data>
   <data name="None" xml:space="preserve">
     <value>None</value>
+    <comment>Function category name</comment>
   </data>
   <data name="Text" xml:space="preserve">
     <value>Text</value>
+    <comment>Function category name</comment>
   </data>
   <data name="Logical" xml:space="preserve">
     <value>Logical</value>
+    <comment>Function category name</comment>
   </data>
   <data name="Table" xml:space="preserve">
     <value>Table</value>
+    <comment>Function category name</comment>
   </data>
   <data name="Behavior" xml:space="preserve">
     <value>Behavior</value>
+    <comment>Function category name</comment>
   </data>
   <data name="DateTime" xml:space="preserve">
-    <value>Date &amp; Time</value>
+    <value>Date and Time</value>
+    <comment>Function category name</comment>
   </data>
   <data name="MathAndStat" xml:space="preserve">
-    <value>Math &amp; Statistical</value>
+    <value>Math and Statistical</value>
+    <comment>Function category name</comment>
   </data>
   <data name="UserDefined" xml:space="preserve">
-    <value>UserDefined</value>
+    <value>Defined by the user</value>
+    <comment>Function category name</comment>
   </data>
   <data name="Information" xml:space="preserve">
     <value>Information</value>
+    <comment>Function category name</comment>
   </data>
   <data name="Color" xml:space="preserve">
     <value>Color</value>
+    <comment>Function category name</comment>
   </data>
   <data name="REST" xml:space="preserve">
     <value>Services</value>
@@ -2434,6 +2444,7 @@
   </data>
   <data name="Component" xml:space="preserve">
     <value>Component</value>
+    <comment>Function category name</comment>
   </data>
   <data name="FunctionReference_Link" xml:space="preserve">
     <value>https://go.microsoft.com/fwlink/?LinkId=722347#</value>
